refactor util.path module into config (these functions are for substituting config variables and so should be handled there)
