+++ lib/spack/spack/package_prefs.py
@@ -12,9 +12,8 @@ from llnl.util.lang import classproperty
 
 import spack.repo
 import spack.error
-from spack.util.path import canonicalize_path
 from spack.version import VersionList
-from spack.config import ConfigError
+from spack.config import ConfigError, canonicalize_path
 
 _lesser_spec_types = {'compiler': spack.spec.CompilerSpec,
                       'version': VersionList}
